Algo #1 added
C1*B1 = C2*B2 = ... = CnBn
EvDiff

+++ DecVarianceProject/Structures/BetInfo.cs
@@ -8,10 +8,10 @@ namespace DecVarianceProject
     public class BetInfo
     {
         public Matches MatchesAndOutcomes { get; set; }
-        public int BetSize { get; set; }
+        public double BetSize { get; set; }
         public double Coef { get; set; }
 
-        public BetInfo(List<int> matchList, List<int> outcomes, int betSize, double coef)
+        public BetInfo(List<int> matchList, List<int> outcomes, double betSize, double coef)
         {
             MatchesAndOutcomes = new Matches();
             this.MatchesAndOutcomes.MatchList = matchList;
